requirements.txt
numpy==1.18.5
streamlit
tensorflow==2.3.1
Pillow==7.2.0
pymongo
pymongo[srv]
requests
dnspython